requirements.txt
boto3
requests
ffmpeg-python
transformers
torch